move pages to app
