+++ src/app/archive/page.js
@@ -1,31 +1,34 @@
+'use client'
+
 import { useEffect, useState } from 'react'
-import { useRouter } from 'next/router'
+import { useRouter, useSearchParams } from 'next/navigation'
 import { format, parseISO } from 'date-fns'
 import useSWRInfinite from 'swr/infinite'
 import useInfiniteScroll from 'react-infinite-scroll-hook'
 import 'intersection-observer' // polyfill for IE11
 
-import jsonFetcher from '../lib/swr/jsonFetcher'
-import { slugify, unslugify } from '../lib/slugify'
+import jsonFetcher from '../../lib/swr/jsonFetcher'
+import { slugify, unslugify } from '../../lib/slugify'
 
 const region = 'us-west-2'
 const s3 = `https://s3-${region}.amazonaws.com/archive.kchungradio.org/`
 
-function ArchivePage() {
+export default function ArchivePage() {
   const router = useRouter()
+  const searchParams = useSearchParams()
   const [searchInput, setSearchInput] = useState('')
   const [search, setSearch] = useState('')
   const [hasNextPage, setHasNextPage] = useState(true)
 
+  const querySearch = searchParams.get('search')
   useEffect(() => {
-    const querySearch = router.query.search
     let search
     if (querySearch) {
       search = unslugify(decodeURIComponent(querySearch))
     }
     setSearchInput(search ?? '')
     setSearch(search ?? '')
-  }, [router.query])
+  }, [querySearch])
 
   const { data, error, setSize, isValidating } = useSWRInfinite(
     (index) => {
@@ -49,7 +52,7 @@ function ArchivePage() {
           setHasNextPage(false)
         }
       },
-    }
+    },
   )
 
   function handleLoadMore() {
@@ -153,5 +156,3 @@ function ArchivePage() {
     </div>
   )
 }
-
-export default ArchivePage
