+++ src/components/Calendar/Calendar.css
@@ -1,0 +1,35 @@
+.calendar {
+  width: 1071px;
+  font-family: courier, sans-serif;
+  background: red;
+  color: white;
+  margin: 0;
+  padding: 0;
+  font-size: 13px;
+}
+.day {
+  display: inline-block;
+  width: 148px;
+  padding-bottom: 1px;
+  text-align: center;
+}
+.blank {
+  display: inline-block;
+  height: 145px;
+  width: 148px;
+}
+a {
+  color: white;
+  cursor: pointer;
+}
+a:hover {
+  background: white;
+  color: red;
+  text-decoration: none;
+}
+a:link {
+  text-decoration: none;
+}
+a:visited {
+  text-decoration: none;
+}
