+++ src/components/Calendar/CalendarCell.js
@@ -1,6 +1,6 @@
 /* eslint-disable react/prop-types */
 
-import React from 'react'
+import './CalendarCell.css'
 
 const formatAmPm = (date) => {
   let hours = new Date(date).getHours()
@@ -32,44 +32,6 @@ export default function CalendarCell({ date, events }) {
           )
         })}
       </div>
-
-      <style jsx>{`
-        .cell {
-          position: relative;
-          display: inline-block;
-          box-sizing: border-box;
-          height: 150px;
-          width: 150px;
-          margin: 0 -2px -5px 0;
-          border: 2px solid white;
-          overflow: hidden;
-        }
-        .date {
-          position: absolute;
-          right: 2px;
-        }
-        .events {
-          position: absolute;
-          overflow: auto;
-          width: 100%;
-          height: 134px;
-          line-height: 1;
-          top: 12px;
-          left: 2px;
-          white-space: nowrap;
-          padding-right: 13px;
-          padding-bottom: 15px;
-          display: flex;
-          flex-direction: column;
-          overflow: hidden;
-        }
-        @media (hover: none), (hover: on-demand) {
-          .events {
-            padding-right: 0px;
-            padding-bottom: 0px;
-          }
-        }
-      `}</style>
     </div>
   )
 }
